introduced macro REAL for switch between float/double (float not supported on older GPUs) added Definitions.h with defined values (like REAL and cell flags) added compiler directive for switching between CPU and GPU

+++ src/inputParser.cpp
@@ -58,7 +58,7 @@ bool InputParser::readParameters
 
 	std::string	buffer;
 	int			i_buffer;
-	double		d_buffer;
+	REAL		d_buffer;
 
 	int line = 0;
 	int numReadValues = 0;
